Update token_ring.adb
Added some stuff to measure the amount of time spent holding the status token. Used an absolute delay to attempt to make the wait time more predictable. Not sure if it worked because there is a much greater range but it is consistently below 0.1. Attempting a different technique and just committing this before I ruin everything.