Update generic_atomic_action_edit.adb
Added a separate task, Sync_Task, to wait a specified period before calling for the atomic action to begin. This then allows the Action_Tasks to stop idleing. Presently however it either doesn't wait long enough to notice that the Action tasks are idle or that they are being blocked by the Sync_Task. Also after the atomic action has been called the Elevator task is the only one to print out, indicating that the other tasks are being starved. TODO: fix starvation after atomic action completed and change sync task to wait longer, or perhaps give it a block function to be called?