Update generator_controllers.adb
This version features a second oscillator to test concurrent input handling in preparation for control testing. Doesn't presently work because I think that the faster timer is interrupting the slower so states don't change. Also whilst the send receive for one of the oscillator pairs works well every so often the receive and transmit lights are out of sync - perhaps interference from the other oscillator?